requirements.txt
eval_type_backport==0.2.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
streamlit==1.32.2
pydantic==2.7.0
anthropic==0.25.0
openai==1.25.0
httpx==0.27.0
click==8.1.7
python-dotenv==1.0.1
pyyaml==6.0.1
requests==2.31.0
pandas==2.2.1
plotly==5.20.0
rich==13.7.1
sqlalchemy==2.0.30
scikit-learn==1.4.2
numpy==1.26.4
sentence-transformers==2.7.0
